Move Pres EB consume links to shared EB normalizer
