+++ src/analysis/retail/evoker/preservation/modules/talents/SparkOfInsight.tsx
@@ -17,9 +17,9 @@ import TalentSpellText from 'parser/ui/TalentSpellText';
 import {
   didEbConsumeSparkProc,
   didSparkProcEssenceBurst,
-  getEssenceBurstConsumeAbility,
   wasEbConsumed,
 } from '../../normalizers/EventLinking/helpers';
+import { getEssenceBurstConsumeAbility } from 'analysis/retail/evoker/shared/modules/normalizers/EssenceBurstCastLinkNormalizer';
 import { ESSENCE_COSTS, MANA_COSTS } from './EssenceBurst';
 
 class SparkOfInsight extends Analyzer {
